Pasalintas Teams failas is komponentu
